instructions/requirements.txt
astroid==2.3.3
attrs==22.2.0
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.0.1
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
distlib==0.3.6
elasticsearch==7.8.0
elasticsearch-dsl==7.2.1
filelock==3.9.0
Flask==2.0.1
fonttools==4.39.3
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.8
isort==4.3.21
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.4
jsonschema==3.2.0
kiwisolver==1.4.4
lazy-object-proxy==1.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
mccabe==0.6.1
mpld3==0.5.9
nltk==3.6.6
numpy==1.24.3
packaging==23.1
pandas==2.0.1
Pillow==9.4.0
platformdirs==3.0.0
psycopg2-binary==2.9.6
pylint==2.4.4
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
regex==2021.8.3
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
text-unidecode==1.3
tqdm==4.64.1
tzdata==2023.3
urllib3==1.21.1
virtualenv==20.19.0
Werkzeug==2.2.2
wrapt==1.11.2
